fix: store prune should not fail if the store contains Node.js (#10193)
close #10131

+++ store/package-store/src/storeController/prune.ts
@@ -75,7 +75,8 @@ export async function prune ({ cacheDir, storeDir }: PruneOptions, removeAlienFi
   let pkgCounter = 0
   await Promise.all(pkgIndexFiles.map(async (pkgIndexFilePath) => {
     const { files: pkgFilesIndex } = await loadJsonFile<PackageFilesIndex>(pkgIndexFilePath)
-    if (removedHashes.has(pkgFilesIndex['package.json'].integrity)) {
+    // TODO: implement prune of Node.js packages, they don't have a package.json file
+    if (pkgFilesIndex['package.json'] && removedHashes.has(pkgFilesIndex['package.json'].integrity)) {
       await fs.unlink(pkgIndexFilePath)
       pkgCounter++
     }
